Load 3.3.706 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@1487 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/object.c
@@ -1597,7 +1597,7 @@ v_function_exists (svalue_t *sp, int num_arg)
 
         default:
             fatal("function_exists(): flags value %ld (from %ld) not implemented.\n"
-                 , flags & ~NAME_HIDDEN, flags);
+                 , (long)(flags & ~NAME_HIDDEN), (long)flags);
             /* NOTREACHED */
         }
     }
